restructure-component-dev
