+++ src/components/shared/reactBit/MetallicPaint.tsx
