+++ src/components/shared/reactBit/SplitText.tsx
